ResultCardViewController class

+++ BuyPolish/Pola/UI/ProductSearch/ScanCode/ProductCard/BPCompanyContentView.h
@@ -9,7 +9,7 @@ typedef enum {
 
 @interface BPCompanyContentView : UIScrollView
 
-@property (nonatomic) int padding;
+@property (nonatomic) CGFloat padding;
 
 @property (nonatomic) CompanyContentType contentType;
